refactor(chains): hoist generic EVM modules monad/ to evm/ (no behavior change)
