+++ src/rumi_protocol_backend/src/chains/evm/adapter.rs
