+++ src/rumi_protocol_backend/src/chains/evm/tests_adapter.rs
