+++ src/rumi_protocol_backend/src/chains/monad/mod.rs
@@ -6,23 +6,17 @@
 //! (Design B): `chain_supplies` and vault debt move only when an on-chain
 //! mint/burn is observed at finality.
 
-pub mod adapter;
-pub mod burn_proof;
 pub mod chain_vault;
 pub mod config;
-pub mod deposit_watch;
-pub mod evm_rpc;
-pub mod hardening;
-pub mod settlement;
-pub mod tecdsa;
-pub mod tx;
+
+// Generic EVM logic now lives in chains::evm; re-export so existing
+// crate::chains::monad::{tx,tecdsa,...} paths keep resolving (mirrors the
+// chains/vault.rs <- chains/monad/chain_vault.rs hoist).
+pub use crate::chains::evm::{adapter, burn_proof, deposit_watch, evm_rpc, hardening, settlement, tecdsa, tx};
 
 pub use adapter::MonadAdapter;
 pub use chain_vault::{ChainVaultStatus, ChainVaultV1};
 pub use config::{monad_default_register_arg, monad_ecdsa_key_name, MONAD_CHAIN_ID, MONAD_ICUSD_DECIMALS};
-
-#[cfg(test)]
-mod tests_adapter;
 
 #[cfg(test)]
 mod tests_config;
@@ -31,28 +25,7 @@ mod tests_config;
 mod tests_chain_vault;
 
 #[cfg(test)]
-mod tests_tecdsa;
-
-#[cfg(test)]
-mod tests_evm_rpc;
-
-#[cfg(test)]
-mod tests_tx;
-
-#[cfg(test)]
-mod tests_deposit_watch;
-
-#[cfg(test)]
-mod tests_settlement;
-
-#[cfg(test)]
-mod tests_hardening;
-
-#[cfg(test)]
 mod tests_open_vault;
 
 #[cfg(test)]
 mod tests_withdraw;
-
-#[cfg(test)]
-mod tests_burn_proof;
